RavenDB-25924 - Smuggler import must not dispose the caller's stream (honor leaveOpen)
On .NET Framework the HTTP stack disposes the request content stream after sending
(StreamContent disposes its inner stream), so Smuggler ImportAsync(Stream) disposed the
caller's stream despite passing leaveOpen: true. On .NET (Core)+ the stream is left open,
so this only manifested on the netstandard2.0 client (e.g. the EmbeddedTests net472 run,
where GZipStream cleanup then wrote to a closed MemoryStream -> ObjectDisposedException).

Wrap the import stream in a non-disposing LeaveOpenStream before handing it to StreamContent,
leaving disposal to the existing DisposeStreamOnce (which only runs for leaveOpen: false).
Unify LeaveOpenStream into Raven.Client.Util (internal) and drop the Raven.Server copy.
