--- src/Raven.Server/Utils/LeaveOpenStream.cs
@@ -1,97 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Threading.Tasks;
-using System.Threading;
-
-namespace Raven.Server.Utils;
-
-public sealed class LeaveOpenStream : Stream
-{
-    private readonly Stream _inner;
-
-    public LeaveOpenStream(Stream inner)
-    {
-        _inner = inner;
-    }
-
-    public override void Flush()
-    {
-        throw new NotSupportedException();
-    }
-
-    public override Task FlushAsync(CancellationToken cancellationToken)
-    {
-        return _inner.FlushAsync(cancellationToken);
-    }
-
-    public override int Read(Span<byte> buffer)
-    {
-        return _inner.Read(buffer);
-    }
-
-    public override int Read(byte[] buffer, int offset, int count)
-    {
-        return _inner.Read(buffer, offset, count);
-    }
-
-    public override Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-    {
-        return _inner.ReadAsync(buffer, offset, count, cancellationToken);
-    }
-
-    public override ValueTask<int> ReadAsync(Memory<byte> buffer, CancellationToken cancellationToken = default)
-    {
-        return _inner.ReadAsync(buffer, cancellationToken);
-    }
-
-    public override long Seek(long offset, SeekOrigin origin)
-    {
-        return _inner.Seek(offset, origin);
-    }
-
-    public override void SetLength(long value)
-    {
-        _inner.SetLength(value);
-    }
-
-    public override void Write(ReadOnlySpan<byte> buffer)
-    {
-        _inner.Write(buffer);
-    }
-
-    public override void Write(byte[] buffer, int offset, int count)
-    {
-        _inner.Write(buffer, offset, count);
-    }
-
-    public override Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-    {
-        return _inner.WriteAsync(buffer, offset, count, cancellationToken);
-    }
-
-    public override ValueTask WriteAsync(ReadOnlyMemory<byte> buffer, CancellationToken cancellationToken = new CancellationToken())
-    {
-        return _inner.WriteAsync(buffer, cancellationToken);
-    }
-
-    public override bool CanRead => _inner.CanRead;
-    public override bool CanSeek => _inner.CanSeek;
-    public override bool CanWrite => _inner.CanWrite;
-    public override long Length => _inner.Length;
-
-    public override long Position
-    {
-        get => _inner.Position;
-
-        set => _inner.Position = value;
-    }
-
-    public override ValueTask DisposeAsync()
-    {
-        return ValueTask.CompletedTask;
-    }
-
-    protected override void Dispose(bool disposing)
-    {
-    }
-}
